Separate react client app and express server
Now React app uses create-react-app
TODO: Authentication is broken for now
