+++ .babelrc
@@ -1,5 +1,5 @@
 {
-  "presets": ["react", "env"],
+  "presets": ["env"],
   "plugins": [
     "transform-async-to-generator",
     "transform-object-rest-spread",
