+++ .env.example
@@ -1,3 +1,5 @@
+NODE_ENV="production" # or development
+
 # database connection
 DB_HOST="host"
 DB_USER="user"
