+++ client/public/401.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <link rel="shortcut icon" type="image/x-icon" href="favicon.ico">
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap.min.css">
+  <title>Brak uprawnień</title>
+</head>
+
+<body>
+  <div class="main container text-center" style="min-height:80vh;display:flex;align-items:center;">
+    <div class="row" style="width:100%;">
+      <div class="col-md-6 col-md-offset-3">
+        <h1 class="display-4 m-b-2 text-center">Brak uprawnień!</h1>
+      </div>
+    </div>
+  </div>
+</body>
+
+</html>
