+++ client/src/components/speakers/preparedDialog.jsx
