+++ client/src/index.js
@@ -3,5 +3,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import './styles/main.css';
+import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
+
+serviceWorker.unregister();
