+++ client/src/reducers/speakersReducer.js
