+++ package.json
@@ -3,67 +3,43 @@
   "version": "2.1.0",
   "description": "Application to schedule lectures",
   "repository": "github:Kyczan/lectures-scheduler",
-  "main": "app.js",
-  "sideEffects": false,
+  "main": "index.js",
   "scripts": {
-    "build": "rimraf ./dist && babel ./src -d ./dist --copy-files && npm run cleanclient && webpack",
-    "build:dev": "npm run build -- -d",
-    "build:prod": "npm run build -- -p && cp ./package* ./dist/",
-    "build:watch": "npm run build:dev -- --watch",
-    "cleanclient": "rimraf ./dist/client/* && cp ./src/client/favicon.ico ./dist/client/",
-    "lint": "esw webpack.config.babel.js src --color",
-    "lint:watch": "npm run lint -- --watch",
-    "nodemon": "npm run build:dev && nodemon -w ./src --exec \"node ./dist/app\"",
-    "start": "npm-run-all --parallel build:watch lint:watch nodemon"
+    "build": "rimraf build && npm run build:server && npm run build:client && cp package* build",
+    "build:client": "cd client && npm run build && cd .. && mkdir build/client && cp -r client/build/* build/client",
+    "build:server": "babel src -d build --copy-files",
+    "prod": "npm run build && NODE_ENV=production node build",
+    "start": "npm-run-all --parallel start:server start:client",
+    "start:server": "nodemon -w src --exec \"babel-node src\"",
+    "start:client": "cd client && npm start"
   },
   "keywords": [],
   "author": "krystian@kowal.pro",
   "license": "ISC",
   "dependencies": {
-    "@material-ui/core": "^1.3.1",
-    "@material-ui/icons": "^2.0.3",
-    "axios": "^0.18.0",
     "body-parser": "^1.18.3",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
     "express-session": "^1.15.6",
     "joi": "^13.3.0",
-    "material-ui-pickers": "^1.0.0-rc.14",
-    "moment": "^2.22.2",
     "passport": "^0.4.0",
     "passport-google-oauth": "^1.0.0",
-    "promise-mysql": "^3.3.1",
-    "pug": "^2.0.3",
-    "react": "^16.4.1",
-    "react-debounce-input": "^3.2.0",
-    "react-dom": "^16.4.1",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^4.3.1",
-    "react-select": "^2.0.0",
-    "redux": "^4.0.0",
-    "redux-thunk": "^2.3.0"
+    "promise-mysql": "^3.3.1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
     "babel-eslint": "^9.0.0",
-    "babel-loader": "^7.1.5",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^1.0.0",
     "eslint": "^5.1.0",
     "eslint-plugin-babel": "^5.1.0",
     "eslint-plugin-react": "^7.10.0",
     "eslint-watch": "^4.0.1",
     "nodemon": "^1.17.5",
     "npm-run-all": "^4.1.3",
-    "rimraf": "^2.6.2",
-    "style-loader": "^0.23.0",
-    "webpack": "^4.15.1",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-cli": "^3.0.8"
+    "rimraf": "^2.6.2"
   }
 }
