--- src/views/401.pug
@@ -1,11 +1,0 @@
-doctype html
-html
-  head
-    link(rel='shortcut icon' type='image/x-icon' href='favicon.ico')
-    link(rel="stylesheet",href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap.min.css")
-    title="Brak uprawnień"
-  body
-    .main.container.text-center(style={'min-height': '80vh', display: 'flex', 'align-items': 'center'})
-      .row(style={width: '100%'})
-        .col-md-6.col-md-offset-3
-          h1.display-4.m-b-2.text-center="Brak uprawnień!"
