--- src/views/index.pug
@@ -1,11 +1,0 @@
-doctype html
-html(lang="pl")
-  head
-    meta(charset="utf-8")
-    meta(name="viewport" content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no")
-    title="Planer wykładów"
-    link(rel='shortcut icon' type='image/x-icon' href='favicon.ico')
-    link(rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500")
-  body
-    div(id="root")
-    script(src="app.dist.js")
